Exclude all CUDA/OpenCL related code when libs not found
When CUDA or OpenCL are not installed on a machine, all the code
that makes use of their corresponding libraries must be excluded
from the binaries, or else compile errors will ensue.

In some cases the libraries are installed, but no compatible GPU
devices are to be found. In those cases, we switch to the CPU
version.